Replace linemin.py with correlated sampling version
